chore(nexus): refactored store module
